test/http: Read full server response before sending next
The test_100_continue test cases sends a series of requests to the test
server with different headers and bodies in order to check how server
responds to it. The problem of this test is that it reads back from
server with single input_stream<>::read() calls. Most of the time it
works OK, but from time to time server sends headers then preempts and
appends the response body later. If this happens client reads response
headers, parses what it wants and proceeds with the next query. But
reading back the response for this next query results in reading back
body from the previous query, thus failing the test.

fixes: #1645

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>