http: add handler_base::verify_mandatory_params()
before this change, this member function was a free function called
by `routes::handle()`, but since `handler_base` is closer to the
`_mandatory_param` and in the following change, we will store
the expected param types in `_mandatory_param`, this makes
`handler_base` a better home for the verification function. so to
prepare for the change, let's move this free function into handler_base.

also, let's mark `_mandatory_param` private, allows us to store the
typing of the params into `_mandatory_param`. as this variable should
not be visible from the outer world. and the typing information should
be opaque from the caller.

Refs #2082
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>