reactor: io_uring: poll first on send/recv
Our speculation mechanism tries to predict if data is available in the
socket (recv) or if there is room in the write buffer (send) and issues
the syscall directly if so. It may make more sense to change it later,
but for now, the expectation is that the request will not be
completed immediately.

Make use of this information by setting the IORING_RECVSEND_POLL_FIRST
flag. This tells io_uring not to try to complete the request immediately,
and instead issue a poll first.