util/concepts: always define SEASTAR_CONCEPT
always define `SEASTAR_CONCEPT` with its parameters and error out
if C++ language support or library support of concept is missing.

since we dropped C++17 support in 01257e60. and C++ concept is one
of the import features introduced by C++20, so we can assume its
presence in the language and the C++ standard library.

in this change, we simplify concepts.hh, it is provided only for
backward compatibility.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>