test/unit/sstring_test: do not test standard library
according to related section of C++ standard draft at $string.ends.with, see
https://eel.is/c++draft/string.ends.with, it is equivalent to
```c++
return basic_string_view(data() + (size() - rlen), rlen) == x;
```
where `rlen` is the smaller of `size()` and `x.size()`.

so, if the parameter equal to the string, this function should return
true.

in existing test of sstring, we are verifying the behavior of standard
library, which is not implemented by us, neither do we change its
behavior. so there is no need to test it.

see also fc86e8cd, which introduced these tests.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>