thread: speed up thread creation in debug mode
In debug mode, we clear the stack to avoid the address sanitizer
thinking we're reading from uninitialized memory (although, typically,
programs write to the stack before reading from it). The problem is
that we use fill_n, which in debug mode doesn't get optimized into
a memset as every write has to pass through asan. This shows up in
profiles as the work to initialize the thread stack can be larger
than the work the thread does.

Fix be using memset() to clear the stack. memset() is intercepted by
asan and is treated as a unit.

Closes #2119