seastar-json2code: extract Parameter class
this change paves the road to #2082. so that we can store more
properties in the Parameter class in this python script and then
translate them into a C++ class which is used to verify and expose
the content and schema of a parameter. so that they can be used
in `handler_base::verify_mandatory_params()` in a follow-up
change.

Refs #2082
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>