core/relabel_config.hh: support metric family config
This patch adds support for metric family config.
A metric family config is a configuration that relates to all the
metrics with the same name (but with different labels).

The configuration would allow changing the labels aggregation for
metrics.

Besides the configuration class, which is similar to
metrics_relabel_config, two enhancements were added to the
relabel_config_regex:

1. empty() which check if it's empty or not.
2. match() that checks if the regex is nonempty and matches a string
without returning the matched group.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>