http/util: add path_decode method
The existing url_decode method is not suitable for decoding path
parameters, as it replaces the + symbol with a space. To help with this,
this commit introduces a new method to decode path parameters
specifically.

Furthermore, the http decoding and encoding utility methods can pass the
const string_view arguments by value instead of reference as the
std::string_view is a cheaply copyable object.