http/request: add get_path_param method
This introduces the get_path_param method to help extract decoded path
parameters out of the http request object.

It lazily calls the path_decode method to decode the path parameters
inside params. I made the choice to lazily decode instead of saving the
decoded parameters into a separate data structure as it was both simpler
to implement and it prevents duplicating the params object in memory.
The alternative I rejected was to introduce a decoded_path_params object
and a perhaps even a parse_path_param() method similar to how
parse_query_param works, although I am willing to be challenged on this.

When the path parameter cannot be decoded successfully, the method
returns an empty string, which is in line with the behaviour of
get_query_param.