build: search protobuf using package config
protobuf v22.0 is packaged by in modern distros, but the FindProtobuf
module shipped along with CMake fails to add the new dependency of
abseil libraries to the linkage. so Seastar fail to link if compiled
with protobuf v22.0 and up.

so, in this change, we try to find protobuf using its CMake package
configuration first, and fall back to the Findprotobuf module. because
find_package() runs in module mode, and fall back to the config mode,
and the protobuf's CMake configuration adds the abseil libraries
dependency correctly, we need to use config mode first.

Refs #2113
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>