tls: add fmt::formatter for tls::subject_alt_name
this change addresses the formatting with fmtlib v10, which stopped
creating the default-generated formatter even if FMT_DEPRECATED_OSTREAM
preprocess macro is defined. so to enable Seastar applications to
print tls::subject_alt_name using {fmt} without defining the formatter
by themselves, we provide the formatter for this class in Seastar.

in which, `fmt::formatter<subject_alt_name_type>` is added for
compatibility with {fmt} < 10, as `format_as()` overloads returning
`std::string_view` can only used as a fallback of `fmt::formatter<>`
since {fmt} v10, before that, we still need a proper specialization
of this template.

Refs https://github.com/scylladb/scylladb/issues/13245

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>