util: add fmt::formatter for bool_class<Tag>
this change addresses the formatting with fmtlib v10, which stopped
creating the default-generated formatter even if FMT_DEPRECATED_OSTREAM
preprocess macro is defined. so to enable Seastar applications to
print bool_class<Tag> using {fmt} without defining the formatter
by themselves, we provide the formatter for this class in Seastar.

the operator<< for bool_class<Tag> is preserved for backward
compatibility. please note, {fmt} format a `bool` variable as "true" or
"false" based on its value, so we don't have to put "true" or "false"
explicitly as we do in the implementation of operator<<.

Refs https://github.com/scylladb/scylladb/issues/13245

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #2144