util/bool_class: use the default-generated comparison operators
now that we use C++20 and up in Seastar, we have the luxury of using
the defaulted equality comparision. since bool_class can be
compared using its member variable `_value`, it is a good candicate of
using this techinique.

in this change, we switch from the hand-crafted equality operators
to the default-generated ones. their behaviors are identical. but
the latter is simpler.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes scylladb/seastar#2145