smp: advise kernel to collapse transparent hugepages eagerly when prefaulting
Instead of deferring merging pages into hugepages to the transparent
hugepage scanner, advise the kernel to do so immediately using the new
MADV_COLLAPSE advise.

Since the advice only takes hold if at least one page is faulted in,
do so in the prefaulter. To reduce the need to do extra work, do it
after exactly one page was prefaulted; this will limit the need to
move pages around if they were not prefaulted as a huge page to begin
with.

Note that prefaulting usually generates huge pages anyway; this makes
the process more reliable in case memory was scrambled.