http: add missing decoding of path part of URL
Seastar's HTTP server implementation correctly does percent decoding
(e.g., "%21" to "!") on the individual pieces of the query part of the
URL (i.e., after the "?"). However, we forgot to do the same decoding
also on the path part of the URL (before the "?"). This patch fixes
this oversight.

The patch fixes two places: The more important place is the code which
splits the URL to the path and the different query parameters. This
is the fix most users will need. The other place is the code which
allows a rule matching a specific path to also match an encoded version.
This patch also adds simple test cases for both.

I also verified that this fix fixes a bug for one user of this HTTP
server (the ScyllaDB project), which did not get correctly-decoded
paths before this patch, and works as expected afterwards.

Fixes #725

Signed-off-by: Nadav Har'El <nyh@scylladb.com>