timed_out_error: add fmt::formatter for timed_out_error
this change is added in the same spirit of cd8a9133d2.

this change addresses the formatting with fmtlib < 10, and without
`FMT_DEPRECATED_OSTREAM` defined. please note, in {fmt} v10 and up,
it defines formatter for classes derived from `std::exception`, so
our formatter is only added when compiled with {fmt} < 10.

please note, unlike the operator<< based formatter of
`operator<<(std::ostream&, const std::exception&)` defined in
log.hh and log.cc, this formatter does not include the exception's
type when formatting an instance of `timed_out_error`.

Refs https://github.com/scylladb/scylladb/issues/13245

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes scylladb/seastar#2163