stall-analyser: move print_command_line_options() out of Graph
this method is not closely related to this class, so let's move
it out. also, by moving it out, we don't need to pass `args` to
`print_graph()` if we don't want to keep `args` as a global
variable. so this change also prepares for the cleanup to
make `args` into `main()` as a local variable. please note,
this changes the order of output, before this change, we print
the command line after the help message explaining how to interpret
the output. after this change, the command line is printed before
the help message.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>