sstring: include <fmt/format.h> instead of <fmt/ostream.h>
<fmt/ostream.h> was include in 2975b612 so we can use ostream_formatter,
but later in bef98ab0, we stopped using ostream_formatter. so this
header is not necessary anymore.

in this change, we trade this include with a smaller dependency.

also, `include <fmt/ostream.h>` in `fair_queue.hh`, where ostream_formatter
is used, but in that header file, we fail to include `<fmt/ostream.hh`.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #2176