src/core/metrics.cc: rewrite set_metric_family_configs
_value_map (the data structure that holds all the registered metrics)
and _metadata, the copy of it that is optimized for metrics reporting
and now independent.

set_metric_family_configs should change them both.

These two passes have little performance impact. It is done only when
set_metric_family_configs, which is something that needs to be explicitly
done externally, is very rare and has no effect afterward.

The performance here is in the order of the number of metric families
which is usually bound to hundreds.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>