util: with_file_input_stream: always close file
The file opened in `with_file_input_stream` is currently
moved to `make_file_input_stream`, but the latter doesn't
close it when the input_stream is closed.

This change turns `with_file_input_stream` into a
coroutine to simplify the close logic, as well
amortize allocations, and then make sure that
the opened file is always closed, on the error
path as well as the error-free path.

Fixes scylladb/seastar#1298

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>

Closes scylladb/seastar#2186