metrics.cc: Safer set_metric_family_configs
It's possible in a corner case that _metadata is null.
For example, if you call set_metric_family_configs before the metadata
was constructed, or theoretically if the metadata didn't completed.

While I couldn't reach this kind of scenario in a unit test I did
managed to reach this in a test.

In this case, we should do nothing, the metadata will be rebuilt with
the updated information.

For safety measure, we set the metadata to dirty() making sure that if
we are somehow in the middle of rebuilding it right now, it will be
rebuilt once again on the next metrics collection.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>