core/future: remove a template parameter from set_callback()
this change is a cleanup.

so we can enforce the type of the second parameter so that it is
always a pointer to `continuation_base<T>`. this save a comment,
and renders the signature of this function more readable.

since this is an internal function, we don't have to preserve
the backward compatibility.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>