core: add label to io_threaded_fallbacks to categorize operations
Adds a label to the `io_threaded_fallbacks` metric that categorizes the
system calls being sent to the thread_pool. This is helpful since any
system call that potentially blocks is sent to the thread pool. And in
cases where the `io_threaded_fallbacks` metric changes it can be
difficult to determine why. With this label though it'll at least give a
user an idea of the type of system call that has caused the change.
Which in turn can quickly allow for possibilities to be eliminated.