iotune: Print file alignments too
When performing measurements, iotune uses IO block size that is not-less
than the DMA alignment reported by file. This in turn is what XFS
reports (spiced with io-properties.yaml numbers, but given it's iotune,
these numbers are not making any difference).

Without knowing those values any reasoning on the produced measurements
is furtune-telling :(

refs: #1698

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>