build: expose ragel generated files as an inteface target
this allows use to attach the properties to this target, so that
we can mark the include directories exposed by it as SYSTEM include
directories.

as ragel could generate headers which does not meet the criteria
enforced by the project, so the compiler may warn and fail the
build like:

```
/home/runner/work/scylladb/scylladb/build/gen/redis/protocol_parser.hh:110:1: error: unannotated fall-through between switch labels [clang-diagnostic-implicit-fallthrough]
  110 | case 1:
      | ^
/home/runner/work/scylladb/scylladb/build/gen/redis/protocol_parser.hh:110:1: note: insert 'FMT_FALLTHROUGH;' to silence this warning
  110 | case 1:
      | ^
      | FMT_FALLTHROUGH;
```

but the seastar application cannot control the behavior of ragel.
if it is a CMake-based project, what it can do is to mark the headers
as SYSTEM headers. but Seastar can make this simpler if the Seastar
project chooses to include Seastar by including its CMakeLists.txt. this
change prepares for this change.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>