github/workflow: do not override <clang++,23,release>
before this change, we

* override the combination of <clang++,23,release>
  by adding settings to build with dpdk enabled,
* override the combination of <clang++,23,debug>
  by adding settings to build with C++20 moduels
  enabled.

but what we intended was to add two additional
builds.

so, in this change, we preserve the existing
combinations of <clang++,23,release> and
<clang++,23,debug> by includeing all the
combinations explicitly, and add the builds
with dpdk and C++20 modules. also the matrix
does not allows one to add extra settings without
overriding an existing combination.
see also https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs

also remove the confusing comment

based on the feedback of reviewer,

> the phrase "only build this combination when dpdk is enable" confusing
> - the phrase "in addition to the above" says the opposite, that we did
> build this combination without dpdk but now we'll build it again with
> dpdk.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #2227