prometheus: revert the condition for enabling aggregation
in d7e16bbca0, we added the query time configuration for disabling
aggregation, but the condition for disabling aggregation was wrong.
before that change, we always enable the aggregation, but after that
change, the aggregation is disabled when the value of query parameter
of "__aggregate__" is not "false". in other words,

- if this query parameter is not set, the aggregation is disabled. this
  changes the existing behavior.
- if this query parameter's value is "false", the aggregation is
  enabeld. this is not the expected behavior.

so, in this change

- change the parameter from `disable_aggregation` to
  `enable_aggregation`. as the positive condition is
  always easier to understand.
- and update the handling accodingly.

Fixes d7e16bbca0d81e2d3b43886c932c242eeffb6e96
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #2232