httpd: mark http_server_control::stop() noexcept
the only callee in `http_server_control::stop()` is
`sharded<Service>::stop()`, which is `noexcept`. so we can assume that
`http_server_control::stop()` is also `noexcept`.

so, in this change, we mark `http_server_control::stop()` noexcept.
this allows us to use `deferred_stop` to stop an `http_server_control`.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes scylladb/seastar#2236