apps/metrics_tester: keep metrics with "private" labels
before this change, only the metrics with labels of {"private": "1"}
are preserved. but if we want to test the metrics aggregation by
name, would be more convenient if we can preserve metrics with the same
name but with different labels.

so, in this change, we relax the criteria so that all metrics with
non-empty "private" labels are preserved.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>