build: add --dpdk-machine option to configure.py
this allows user and our toolings to specify the target architecture
when building Seastar. in particular, it prepares us to avoid using
the default "-march=native" setting when building Seastar with DPDK
enabled, which could fail the build when building with a compiler
supporting AVX512, while the building host does not.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>