github: extract test.yaml into a resusable workflow
github's matrix machinery does not allow us to attach extra elements to
an existing matrix, for instance, we cannot add the combination of
`{compiler: clang-18, standard: 23, mode: release, enables: dpdk}` to a
matrix of `{compiler: [clang-18, gcc-13], mode: [debug, release], standard: [20, 23]}`,
without overwriting the combination of
`{compiler: clang-18, standard: 23, mode: release}` generated by the
matrix above. as the combination with "enables: dpdk" matches with
it.

to address this issue, we use the approach that we were using
when defining the circleci workflow: to define a parameterized workflow,
so that we can customize the behavior of the caller workflow
using either a matrix or a hardwired combination.

this change just extract the reuseable workflow out, and call it
in the parent workflow. we will replace the unrolled `inputs`
list with 3 different jobs, with which we will have

- 12 combinations generated with the matrix
- 1 combination for dpdk build
- 1 combination for C++20 build

instead of:

- 10 combinations
- 1 combination for dpdk build
- 1 combination for C++20 build

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>