github: add 2 testing combinations back to the matrix
this change adds the combinations of

- {compiler: clang++-18, standard: 23, mode: release}
- {compiler: clang++-18, standard: 23, mode: debug}

back.

before this change, the `include` keyword instructs
github workflow to override these two combinations with
the ones which builds with dpdk and with C++20 modules.
this hurts the coverage of the test matrix.

after this change, instead of using `include` keyword,
we just add two more jobs which are dedicated for
testing these two combinations without overriding
the existing ones.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>