treewide: include required header files
before this change, when building with C++20 modules enabled, some
symbols are not accessible, when compiling source files like
src/core/resource.cc if some features are enabled using macros like
SEASTAR_HAVE_HWLOC and SEASTAR_HAVE_URING.

in this change, we include the missing headers or move the include
precessor directives, so that the .cc files can have access to the
used declaration of the used symbols.

please note, this change not only enables us to build C++20 modules
with SEASTAR_HAVE_HWLOC and/or SEASTAR_HAVE_URING defined, it is also
necessary also from the correctness point of view.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>