sstring: deprecate formatters for vector and unordered_map
Seastar is an event-driven framework, not a collection of libraries
for multiple purposes. also, when migrating to a {fmt} only formatting
solution, the operator<< based printers can actually makes our lives
more difficult. for instance, Boost.test expects operator<< and fall
back to `boost_test_print_type()`, so even if we provide a templated
`boost_test_print_type()` which accepts all types which can be formatted
with {fmt}, if Boost.test is able to find the operator<<-based
formatter for a std::vector, it just picks it, and then hits the brick
wall, as the elements in the vector does not provide an operator<<-based
formatter.

instead of enabling these operator<<:s to print the element with
fmt::formatter support, let's just disable them on user's request.

in this change, a new option is added. so that user can disable these
formatter on request. and these two formatters are marked deprecated.
so in future, we can remove them when we bump up the API level or just
completely drop them.

Fixes #1544
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>