github: cancel a job if it takes longer than 40 minutes
in general, it takes less than 34 minutes to build the tree from
scratch and to test all tests, even with DPDK enabled. but to be
on the safe side, let's add 6 more minutes.

this helps to stop if we run into some unexpected infra issues,
or to identify some very time consuming tests.

see also
https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idtimeout-minutes

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>